+++ packages/flutter_adaptive_scaffold/CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ## 0.1.12
 
-* Add breakpoints for Expanded and Extra Large.
+* Add breakpoints for mediumLarge and extraLarge.
 * Add height based breakpoint checks
 
 ## 0.1.11+1